+++ class.mesh-upgrades.php
@@ -59,6 +59,10 @@ class Mesh_Upgrades {
 		if ( version_compare( $GLOBALS['mesh_current_version'], '1.1.6', '<' ) ) {
 			$this->version_1_1_6();
 		}
+
+		if ( version_compare( $GLOBALS['mesh_current_version'], '1.1.7', '<' ) ) {
+			$this->version_1_1_7();
+		}
 	}
 
 	/**
@@ -144,6 +148,14 @@ class Mesh_Upgrades {
 	function version_1_1_6() {
 		update_option( 'mesh_version', '1.1.6' );
 		$GLOBALS['mesh_current_version'] = '1.1.6';
+	}
+
+	/**
+	 * Nothing to update here, just the version to 1.1.6
+	 */
+	function version_1_1_7() {
+		update_option( 'mesh_version', '1.1.7' );
+		$GLOBALS['mesh_current_version'] = '1.1.7';
 	}
 }