+++ class.mesh.php
@@ -96,9 +96,6 @@ class Mesh {
 
 		add_filter( 'get_edit_post_link', array( $this, 'get_edit_post_link' ), 10, 3 );
 
-		// Exclude certain taxonomies for WordPress SEO / Yoast.
-		add_filter( 'wpseo_sitemap_exclude_taxonomy', array( $this, 'wpseo_sitemap_exclude_taxonomy' ) );
-
 		// @since 1.1.3
 		add_action( 'after_setup_theme', array( $this, 'after_theme_setup' ) );
 	}
@@ -221,6 +218,7 @@ class Mesh {
 		$in['paste_strip_class_attributes'] = 'none';
 		$in['paste_text_use_dialog']        = true;
 		$in['wpeditimage_disable_captions'] = true;
+		$in['allow_script_urls']            = true;
 		$in['plugins']                      = 'tabfocus,paste,media,wordpress,wpgallery,wplink';
 
 		// Only add in our editor styles if we have the file.
@@ -447,6 +445,29 @@ class Mesh {
 				update_post_meta( $section->ID, '_mesh_template', $template );
 			}
 
+			/**
+			 * Process Section Meta
+			 */
+			$default_section_meta = array(
+				'css_class',
+				'lp_equal',
+				'title_display',
+				'push_pull',
+				'collapse',
+				'blocks',
+				'post_title',
+				'post_status',
+				'template',
+				'menu_order',
+				'featured_image',
+			);
+
+			/**
+			 * This filter is used to remove or add elements to the default section meta
+			 * @todo "meta" related to a section
+			 */
+			$default_section_meta = apply_filters( 'mesh_default_section_meta_fields', $default_section_meta );
+
 			// Save CSS Classes.
 			$css_classes           = explode( ' ', $section_data['css_class'] );
 			$sanitized_css_classes = array();
@@ -461,6 +482,14 @@ class Mesh {
 				delete_post_meta( $section->ID, '_mesh_css_class' );
 			} else {
 				update_post_meta( $section->ID, '_mesh_css_class', $sanitized_css_classes );
+			}
+
+			$featured_image = $section_data['featured_image'];
+
+			if ( empty( $featured_image ) ) {
+				delete_post_meta( $section->ID, '_thumbnail_id' );
+			} else {
+				update_post_meta( $section->ID, '_thumbnail_id', (int) $featured_image );
 			}
 
 			// Save LP Equal.
@@ -494,6 +523,21 @@ class Mesh {
 				delete_post_meta( $section->ID, '_mesh_collapse' );
 			} else {
 				update_post_meta( $section->ID, '_mesh_collapse', $section_data['collapse'] );
+			}
+
+			// Process our custom meta
+			foreach( $section_data as $data_key => $data_field ) {
+				// Do not process default keys
+				if( in_array( $data_key, $default_section_meta ) ) {
+					continue;
+				}
+
+				// Save Custom Meta Field.
+				if ( empty( $section_data[ $data_key ] ) ) {
+					delete_post_meta( $section->ID, '_mesh_' . $data_key );
+				} else {
+					update_post_meta( $section->ID, '_mesh_' . $data_key, $section_data[ $data_key ] );
+				}
 			}
 
 			// Process the section's blocks.
@@ -780,6 +824,11 @@ class Mesh {
 			return;
 		}
 
+		// Do not show content on password protected posts
+		if ( post_password_required() ) {
+			return;
+		}
+
 		$sections = mesh_display_sections( $wp_query->post->ID, false );
 
 		echo apply_filters( 'mesh_loop_end', $sections, $wp_query->post->ID ); // WPCS: ok.
@@ -997,30 +1046,6 @@ class Mesh {
 			'p'      => array(),
 			'br'     => array(),
 		);
-	}
-
-	/**
-	 * Filter to exclude the taxonomy from the XML sitemap.
-	 *
-	 * @since 1.1.3
-	 *
-	 * @param boolean $exclude       Defaults to false.
-	 * @param string  $taxonomy_name Name of the taxonomy to exclude.
-	 *
-	 * @return boolean
-	 */
-	public function wpseo_sitemap_exclude_taxonomy( $exclude = false, $taxonomy_name ) {
-
-		$excluded_taxonomies = array(
-			'mesh_template_usage',
-			'mesh_template_types',
-		);
-
-		if ( in_array( $taxonomy_name, $excluded_taxonomies ) ) {
-			return true;
-		}
-
-		return false;
 	}
 
 	/**