+++ mesh.php
@@ -3,7 +3,7 @@
  * Plugin Name: Mesh
  * Plugin URI: http://linchpin.agency/wordpress-plugins/mesh
  * Description: Adds multiple sections for content on a post by post basis. Mesh also has settings to enable it for specific post types
- * Version: 1.1.6
+ * Version: 1.1.7
  * Text Domain: mesh
  * Domain Path: /languages
  * Author: Linchpin
@@ -21,7 +21,7 @@ if ( ! function_exists( 'add_action' ) ) {
 /**
  * Define all globals.
  */
-define( 'LINCHPIN_MESH_VERSION', '1.1.6' );
+define( 'LINCHPIN_MESH_VERSION', '1.1.7' );
 define( 'LINCHPIN_MESH_PLUGIN_NAME', __( 'Mesh', 'mesh' ) );
 define( 'LINCHPIN_MESH__MINIMUM_WP_VERSION', '4.0' );
 define( 'LINCHPIN_MESH___PLUGIN_URL', plugin_dir_url( __FILE__ ) );
@@ -37,11 +37,13 @@ define( 'LINCHPIN_MESH_DEBUG_MODE', false );
 $GLOBALS['mesh_current_version'] = get_option( 'mesh_version', '0.0' ); // Get our current Mesh Version.
 
 include_once 'class.mesh-settings.php';
+include_once 'class.mesh-controls.php';
 include_once 'class.mesh-templates.php';
 include_once 'class.mesh-pointers.php';
 include_once 'class.mesh.php';
 include_once 'class.mesh-upgrades.php';
 include_once 'class.mesh-reponsive-grid.php';
+include_once 'class.mesh-integrations.php';
 
 $mesh          = new Mesh();
 $mesh_pointers = new Mesh_Admin_Pointers();
@@ -262,7 +264,6 @@ function mesh_add_section_admin_markup( $section, $closed = false, $return = fal
 	// It's important to pass this to the admin to control if a
 	// section's blocks have a post_status of publish or draft.
 	$block_count       = $templates[ $selected_template ]['blocks'];
-
 	$css_class         = get_post_meta( $section->ID, '_mesh_css_class', true );
 	$lp_equal          = get_post_meta( $section->ID, '_mesh_lp_equal', true );
 	$offset            = get_post_meta( $section->ID, '_mesh_offset', true );
@@ -293,7 +294,14 @@ function mesh_add_section_admin_markup( $section, $closed = false, $return = fal
  *
  * @return array|WP_Query
  */
-function mesh_get_sections( $post_id, $return_type = 'array', $statuses = array( 'publish' ) ) {
+function mesh_get_sections( $post_id = '', $return_type = 'array', $statuses = array( 'publish' ) ) {
+
+	// If no Post ID fall back to the current global ID
+	if( empty( $post_id ) ) {
+		global $post;
+		$post_id = $post->ID;
+	}
+
 	$args = array(
 		'post_type'      => 'mesh_section',
 		'posts_per_page' => apply_filters( 'mesh_templates_per_page', 50 ),
@@ -447,24 +455,27 @@ function mesh_display_sections( $post_id = '', $echo = true ) {
  *
  * @param  int    $section_id  Post ID of the target Section.
  * @param  string $post_status Post Status of the target Section.
+ * @param  int    $number_needed
  *
  * @return array
  */
-function mesh_get_section_blocks( $section_id, $post_status = 'publish' ) {
+function mesh_get_section_blocks( $section_id, $post_status = 'publish', $number_needed = 50 ) {
+
 	$args = array(
         'post_type' => 'mesh_section',
         'post_status' => $post_status,
-        'posts_per_page' => 50,
+        'posts_per_page' => $number_needed,
         'orderby' => 'menu_order',
         'order' => 'ASC',
         'post_parent' => (int) $section_id,
     );
 
-    if ( isset($_GET['preview_id']) && isset($_GET['preview_nonce']) ) {
+    if ( isset( $_GET['preview_id'] ) && isset( $_GET['preview_nonce'] ) ) {
         $id = (int) $_GET['preview_id'];
 
-        if ( false === wp_verify_nonce( $_GET['preview_nonce'], 'post_preview_' . $id ) )
-            wp_die( __( 'Sorry, you are not allowed to preview drafts.', 'mesh' ) );
+        if ( false === wp_verify_nonce( $_GET['preview_nonce'], 'post_preview_' . $id ) ) {
+	        wp_die( __( 'Sorry, you are not allowed to preview drafts.', 'mesh' ) );
+        }
 
         // Make sure $post_status is an array
         $args['post_status'] = is_array( $args['post_status'] ) ? $args['post_status'] : array( $args['post_status'] );
@@ -588,7 +599,7 @@ function mesh_maybe_create_section_blocks( $section, $number_needed = 0 ) {
 		}
     }
 
-	return mesh_get_section_blocks( $section->ID, array( 'publish', 'draft' ) );
+	return mesh_get_section_blocks( $section->ID, array( 'publish', 'draft' ), $number_needed );
 }
 
 /**